refactor: create payment and recipient packages inside model/order package
