+++ src/main/java/com/kopchak/worldoftoys/model/order/OrderDetails.java
@@ -1,8 +1,11 @@
 package com.kopchak.worldoftoys.model.order;
 
+import com.kopchak.worldoftoys.model.order.payment.Payment;
+import com.kopchak.worldoftoys.model.order.recipient.OrderRecipient;
 import com.kopchak.worldoftoys.model.product.Product;
 import com.kopchak.worldoftoys.model.user.AppUser;
 import jakarta.persistence.*;
+import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -23,6 +26,7 @@ public class OrderDetails {
 
     @Column(nullable = false)
     @Temporal(TemporalType.TIMESTAMP)
+    @NotNull(message = "Invalid data and time: data and time is mandatory")
     private LocalDateTime dateTime;
 
     @Column(nullable = false)
