+++ src/main/java/com/kopchak/worldoftoys/model/order/recipient/number/PhoneNumber.java
@@ -1,6 +1,6 @@
-package com.kopchak.worldoftoys.model.order;
+package com.kopchak.worldoftoys.model.order.recipient.number;
 
-import com.kopchak.worldoftoys.model.product.Product;
+import com.kopchak.worldoftoys.model.order.recipient.OrderRecipient;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Pattern;
@@ -8,8 +8,6 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-
-import java.util.Set;
 
 @Entity
 @AllArgsConstructor
@@ -27,12 +25,12 @@ public class PhoneNumber {
 
     @Column(length = 2, nullable = false)
     @NotBlank(message = "Invalid operator code: code is mandatory")
-    @Pattern(regexp = "\\d{2}", message = "Invalid operator code: code format is invalid")
+    @Pattern(regexp = "\\d{2}", message = "Invalid operator code: code '${validatedValue}' format is invalid")
     private String operatorCode;
 
     @Column(length = 7, nullable = false)
     @NotBlank(message = "Invalid number: number is mandatory")
-    @Pattern(regexp = "\\d{7}", message = "Invalid number: number format is invalid")
+    @Pattern(regexp = "\\d{7}", message = "Invalid number: number '${validatedValue}' format is invalid")
     private String number;
 
     @OneToOne(mappedBy = "phoneNumber")
